population-dist: create function to plot dist
This function tries to replicate Figure 1a from "Importance of being
uncertain" with the text annotations of the mean and standard deviation,
but also the arrows and rectangle shading.

This function makes it easy to experiment with different means, standard
deviations, and population sizes to see the changes in the density plot.